Rewrite links to Developer Docs rather than GitHub where possible
For a few repositories - and hopefully soon all repositories - we
import the `docs/` folder into the Developer Docs. Where imported
docs have relative links to `.md` files in the `docs/` folder, it
is a link to a doc that has been imported to the Developer Docs.
In these cases, we want to rewrite the link to point to the
Developer Docs version, rather than the GitHub version.

Relative docs can be referenced with an explicit current directory
prefix (`./relative.md`) or without (`relative.md`), so I've added
tests for both cases.

We also need to be mindful of the current context, i.e. if the
document we're parsing is the top-level `README.md`, then a path
to the docs folder will look like `docs/foo.md`. But if we're
parsing a docs document, then a path to the docs folder will be
simply `foo.md` (we're already in the directory). Again, I've
added tests for this.

Many repos currently have a `docs` folder but haven't opted in to
having that folder ingested. Again, I've written a test that this
URL rewriting should only apply to repos whose `consume_docs_folder`
property is `true`.

Finally, we currently only import top-level documents under the
`docs` folder, so I've added an explicit test that subfolder
documents aren't rewritten to Developer Docs URLs.