Clarify the what the stack is for terraform commands
When looking to run terraform commands people may be tempted to use the
`govuk` stack for apps that are within the `govuk` space. However this
is not correct.

This update adds the information about what each stack is usually used
for next to the description of the possible values so it is easier to
spot.