Document how to manually redirect routes in app console
It is crucial that we know how to manually redirect routes if
we run into issues. Prior to this commit, this knowledge wasn't
particularly well documented anywhere.

Rather than list all the different scenarios in which we want
to redirect and making the reader scan the page for the option
relevant to them, I've attempted to provide some brief context
at the top of the document, followed by a flow chart of sorts.
The first heading, "Choosing a redirect strategy", describes the
thought processes we should follow before applying any redirects.

I've documented the preferred approach for manual redirects, as
well as warn the reader that it won't work for Short URL Manager
routes (something I discovered by accident while trying the code
out!). And, in a recent incident, we were forced to apply
redirects at the Router API level, which is also a useful thing
to document.