Document Content Store / Router API comms edge case
There was recently a case where a redirect had been removed, and
was showing as such in the Content Store, but the redirect was
still happening. It turns out the route registration had failed
because of reprovisioning of the Router machines.

This same situation could easily happen in reverse: a redirect
being ADDED, but failing to make its way into Router API. Rather
than semi-polluting the content item in Publishing API to force
through the redirect, it's better in cases such as these to just
give Content Store a bit of a kick, as it's already in the state
we want it to be in. Router API just needs to catch up.