Allow for non-default ports in URLs
In an earlier iteration of this code, the build was breaking
while parsing README for govuk-dependencies:
https://github.com/alphagov/govuk-dependencies/commit/07b67374645958591829dca5539abcdfd539a48e#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R32

We should have a test case that checks for non-standard ports in
URLs and ensures we can parse them correctly (as absolute URLs
rather than relative URLs). Adding this test gives us confidence
that the next few commits don't break anything.

NB whilst this should probably have a `http://` prefix in the README
itself, we don't want govuk-developer-docs to be brittle to poor
formatting in other repositories.