Rewrite "Load test Email Alert API"
This documentation relied on an approach that is no longer supported by
Email Alert API (DelayProvider) and rake tasks that are deleted. This
documentation has been updated to point people towards the code that
used to exist so that this can be re-used.

I decided to remove references to testing Notify load as I felt that
shouldn't be advised, it seemed more natural to not mention it than to
say "Don't do it, but here is how you do it".

I also removed the puppet disable steps as it isn't clear that someone
would need these for any future load testing.