Update "Receiving email" documentation for Google groups
The process for testing email receiving in Email Alert API is now simpler
due to the use of Google groups. This updates the documentation to
reflect this and tries to be more helpful.

It also updates the capitalisation of Integration and Staging for places
where they're used as a noun.