Remove the digest steps documentation
This documentation is rather tangential to the topic of receiving email
and provides some very specific steps that are a little dangerous
(deleting items from the DB) and at risk of aging badly (coupled to
application code).

My expectation is that this documentation is so infrequently needed that
it has barely been used. I would expect that, should a wider need for this to
emerge, then a rake task would be introduced into the application to
achieve it.