Add a link to documentation on receiving Email Alert API emails
This existing documentation doesn't apply to Email Alert API, so I've
added this link to avoid people not understanding it's relevance. I'm
not convinced the remaining documentation is fully accurate (Content
Publisher uses Google groups for instance) or clear that it's for
non-live environments. However, this is a problem outside the scope
of my current task.