Migrate parts of Email "troubleshooting" to "How they work"
This migrates the somewhat distinct part of troubleshooting that is not
replaced elsewhere, courtesy copies, to the "How they work" document. I
felt that this also needed a contextual introduction to the types of
email to explain what courtesy copies to expect.

There didn't seem to be further need for the troubleshooting content as
it's replicated elsewhere.