Improve the Q&A format at top of doc & apply some review feedback
It should be simple for developers to see at-a-glance the most
relevant redirect option for them. I've attempted to do this with
an FAQ type section at the top, linking off to specific subsections.

We continue to discourage developers from updating Router API
directly, but in an incident it's worth highlighting how to do
that quickly; we'd then expect the developer to update through the
publishing platform to keep it all in sync.