Use native 'ruby/setup-ruby' options for caching
'setup-ruby' can run 'bundle install' for us, and cache the result.
https://github.com/ruby/setup-ruby#caching-bundle-install-automatically

Using the 'bundler-cache: true' option in 'ruby/setup-ruby'
removes our dependency on 'actions/cache' and is much more
readable config.

After the initial build [1], it leads to build times of just under
4 minutes [2], which is comparable to what we have on master
already [3]. So this is just about readability rather than
improving the build time.

[1]: https://github.com/alphagov/govuk-developer-docs/runs/1589338635
[2]: https://github.com/alphagov/govuk-developer-docs/runs/1589391081
[3]: https://github.com/alphagov/govuk-developer-docs/runs/1588704546

Thanks to @tomnatt for the tip:
https://tomnatt.blogspot.com/2020/12/a-first-look-at-github-actions.html