Use default GitHub 'actions/checkout' depth when cloning
According to https://github.com/actions/checkout, the default
depth is 1, meaning only 1 commit (the latest one) gets
downloaded when cloning.

This is all that's needed for the purposes of running the tests.
`fetch-depth: 0` (which is an infinite depth) is needed for the
actual _deployment_ so that the correct 'Last Updated' dates
are fetched for each local document.

Removing the infinite fetch depth for the initial cloning has
a significant improvement on the build time, from around 3m50s
to around 2m20s. See example build here:
https://github.com/alphagov/govuk-developer-docs/runs/1589463392