Stop returning early when encountering empty docs folders
Some apps have no `docs/` folder, in which case GitHubRepoFetcher
returns `nil`. When ProxyPages encountered this `nil`, rather than
just returning an empty array `[]` to the mapped application, it
returned _early_ so that the overall return value for the `api_docs`
method was `[]`, clearing any legitimate app doc values that came
before it.