Use application template for rendering external docs
This commit has apps' docs' links appear as sub menu links in
the sidebar, and renders the 'last updated' banner conditionally.
It needs to be conditional as we don't want to render the banner
on top-level app pages (e.g. /apps/content-store.html) or on the
'Application ownership by team' page.