Re-slug `/apis/**` to `/apps/**`
We'll soon be pulling in docs from all repositories, whether they
are APIs or not. The old link structure therefore no longer makes
sense.

This commit re-slugs imported documents like so:

```diff
  /apps/publishing-api.html
- /apis/publishing-api/dependency-resolution.html
+ /apps/publishing-api/dependency-resolution.html
 ```

This is a more RESTful URL structure.

It also sets up redirects for all existing imported docs. I expect
we'll be able to remove these redirects in a few weeks/months when
we can be sure nothing is linking to them.