Only attempt to loop over imported docs if not `nil`
Now that we import _all_ repositories' `docs/` folders, we've run
into an issue whereby any repo without a `docs/` folder returns
`nil`, and we can't call `.each` on `nil`.

This probably indicates that our interface could be improved a bit,
although it is useful to be able to query whether there are any
docs by checking if `nil` rather than counting how many docs are
returned and then concluding that there aren't any docs after all
if 0 is returned. That said, we'll be changing the UI in subsequent
commits, so I'm not too worried about improving the underlying
code just yet.