Rename 'AppDocs.pages' to 'AppDocs.apps'
The method does actually return an array of elements of type `App`,
so `pages` is quite a disingenuous name, all the more problematic
now that we have a concept of imported app 'docs' (synonymous with
pages).

Renaming 'pages' to 'apps' makes it clearer what object types we
are dealing with.