WIP: Call `assets:precompile` directly, instead of `build`
This prevents us from clearing the cache, which _hugely_ speeds up
being able to run a full test suite as we no longer have to fetch
all the docs from GitHub again.

It is safe to do so here because the actual deployment calls
`build` directly (https://github.com/alphagov/govuk-developer-docs/blob/55ffdc5e34cd1a3168f31f7fa3e29a5e803fec3b/.github/workflows/ci.yml#L52)
which does a cache clear before the build. So deployments are
unaffected. This change only makes life easier locally.

HOWEVER - this is a WIP because many tests rely on clearing the
cache and putting arbitrary things in it. So we need to rewrite
those tests to mock the cache instead.

I've had to expose a `cache=` method to allow tests to write a
different cache object. Ideally we shouldn't edit application code
for the sake of our tests, but as this isn't an API people are
using then I don't see much of an issue.

I've removed the `@all_alphagov_repos ||=` cache, which was making
it difficult to write tests. It also wasn't very effective as an
optimisation tool, as it was only saving a (very cheap) lookup in
the cache, rather than say network requests.