Move 'private_repo?' check into the 'readme'/'docs' methods
We're already checking if a repo is private in the
'consume_docs_folder' method in AppDocs, but this will be removed
in the next commit. It's safer to move the check to a lower level
in the code, to ensure that there's no possible way we can
accidentally retrieve private information through these methods.

I had to stop using `app_name` and start using `github_repo_name`
for `govuk-content-store-examples`, which is actually called
`govuk-content-schemas` on GitHub.

I also had to explicitly reject private repos from ProxyPages
and api_layout, as otherwise the build would error with
`RuntimeError: alphagov/govuk-knowledge-graph not found` (which
is a private repository). Hopefully we can continue to refactor
govuk-developer-docs to be less brittle around this.