Make DocumentTypes/DocumentTypesCsv tests more resilient in runs
These tests would often fail if the cache exists already (from an
earlier test run or build). In the case of document_types_spec.rb,
the keys would sometimes return in a different order, so I've
now arranged them alphabetically before asserting them. And in the
case of the CSV tests, I've bypassed having to worry about the
cache and simply mock the method calls rather than stub the
network requests. This has led to easier to read tests!