Rewrite the main `GitHubRepoFetcher.instance.docs` test
There's quite a lot of mocking going on, but I think this is an
improvement on what we had previously (you no longer need to
scroll in order to read the entire test, for a start!).

This test is mainly about how the responses from network requests
and method calls are combined into an array result, so rather than
do all the complicated setup to get those method calls returning
something we can test against, it's better to simply mock the
return values of those method calls.