Update frontend linting approaches for standardx
This updates the linting guidance to reflect the standardx is the
approach used by GOV.UK for JS linting typically.

Rather than going specific about the linting recommendations I've added
links to the GDS way as that explains some of the nuances of linting
tools and saves us repeating this in our docs.