Rename AppDocs to Applications
Now that we have a concept of 'documentation from a given app's
repo', the term 'AppDocs' is misleading - it suggests that it is
that. When actually it is a collection of what we now call `App`s
(as per the class in `app.rb`), and docs are separate. To attempt
to remove the confusion, I've renamed the class. I'll rename some
of its methods next.