Small improvements to intro to Docker
Rather than have to keep this up to date with the latest docker version
or risk a user being confused if their terminal shows a different
version to whats in the docs, I've replaced these with placeholders and
a link to find the latest version of Docker Engine.

Add a note about how to exit the docker terminal

Most users will likely have the CLI-fu to do this, but on the off chance
a user might feel "trapped" inside this shell-within-a-shell, I've added
a short note about the `exit` command.