New app pages
Problem: application pages are unnecessarily long, requiring the
reader to scroll to get to commonly used, important parts of the
page, e.g. GitHub repo link.

Hypothesis: we can cleanly display the most useful at-a-glance
information above the fold. The reader can scroll to dig further
if they wish.

Solution: use of govuk_publishing_components that are designed to
present lots of information in a small amount of screen real
estate. Two components were considered:

- [Summary list](https://components.publishing.service.gov.uk/component-guide/summary_list)
- [Metadata](https://components.publishing.service.gov.uk/component-guide/metadata)

The former was designed to be displayed at the end of a user's
form journey, to replay their answers back to them. It extends the
[Summary List component in the Design System](https://design-system.service.gov.uk/components/summary-list/).
It has since been reused to show key information at the top of
admin pages in Collections Publisher.

The latter was designed to display publish dates and organisation
tags on pages like Manual Frontend pages. It uses less space than
the former.

I've taken the opportunity to make each subsection more terse.