Add `continuously_deployed` property to API payload
This should have been done in #2949 - I wrongly assumed the
application.yml contents would automatically be consumed in the
apps.json endpoint. It turns out it has to be explicitly added to
the api_payload method.

This commit adds a `continuously_deployed` method to retrieve the
property of the same name from the config, defaulting to `false`
if it doesn't exist. I've added tests for this, as well as a
missing text for the `api_payload` method.