Add "govuk-link" class and use List component where possible
I had to bump the version of govuk_publishing_components used, as
the previous version didn't have a List template available. The
old version, v21.5.0, was from October 2019! Looking at the diff
between 21.5.0 and 21.5.1, it's probably the Rails 6 upgrade that
caused a conflict somewhere:
https://github.com/alphagov/govuk_publishing_components/compare/1e8462fadf7c2f4430856bb3fc6cf34ddde08e96..4cfb2cfe0909f017a99cef3c80a11d1be070505e

Specifying a minimum dependency of v23.0 and forcing a general
update with `bundle update` seems to have done the trick.

I attempted to get `_links.html.erb` using the List component too,
but the styling wasn't applied as it is pulled into the Summary
List component, so have left this as-is (but added the 'govuk-link'
classes).