Bring asset documentation up to date
This is updated to reflect that static assets are now served from
www.gov.uk and that uploaded assets aren't served from a share anymore.

I also decided to include the peculiar other thing hosted on
assets.publishing.service.gov.uk since it is rather a surprise to find
that one of our apps is proxied there as only a service to development
versions of GOV.UK. Hopefully I've explained it ok as it's a rather
confusing concept.