Explain that new repositories require github About
During setup of accounts-api we found CI errors and a very obscure stack
trace. It turns out the code is currently set up to expect that each
repo has an "About section". Without it the application listing renders
fine but shows up empty.

However the application page fails to render and there is an error
message when building pages with Middleman.

This was quite confusing, so until we can fix the code to fail more
gracefully or with a better warning, I have added a note for the next
folks who try and set up a repo.