Add tmp/ dir
We'll use this in the next commit for unit testing, but we
want to git-ignore the directory. We can't do a .keep and a
.gitignore in the same commit, otherwise the directory won't ever
be created.

I've decided to add this to the top level directory, rather than
inside `spec/` (where the test that needs a temporary directory
lives), as this follows Rails convention of creating a `tmp`
directory at the root when running `rails new`. See
https://github.com/alphagov/content-publisher/commit/4ba770c5b0e31023a652a41fdd5360c55084011d#diff-cf10d3eb4b80f1fdd74306ab6e6152f1822b19451b959eba448ba2d0b2beb22b

For this reason I've also named the file `.keep` rather than the
other popular alternative, `.gitkeep`.