Fix #2875 - local uncommitted files causing build errors
When creating a file locally, before it has been committed to
version control, if we attempt to build and view the Developer
Docs we see errors such as
`NoMethodError: undefined method `author_date' for nil:NilClass`

This is because the code tries to find the last commit associated
with the file, in order to render the 'Last updated' banner. There
is no commit yet, so it errors.

To avoid complicating the 'last_updated' partial with a conditional,
I've gone down the route of passing a temporary link: today's date is
used as the last updated date, and the link is to nowhere ("#").

This is only an issue on the developer's local machine while
they're developing. It is never an issue on production or on CI,
by which time all files in the build have been committed.

Trello: https://trello.com/c/wrC4T1cr/220-creating-or-moving-a-page-gives-a-nil-parsing-error