Remove fab tasks from "ntp drift too high" page
The first task reports the NTP status to icinga, but that seems
redundant if there's already an alert for it; the second resyncs the
clock.

If the clock is actually fine, and the problem was with reporting the
icinga check, then the clock sync will take <1 second and hopefully
the next check will succeed.  So there's no harm in doing a redundant
sync.