Document some gotchas with role names when deploying Terraform.
`gds-cli` has some usability issues in this area which have caused
confusion recently (to another user and to me when troubleshooting the
issue).

For now, document the issue so that at least it's easier to discover
that there are similar, easily-confused role names and where to find the
canonical source of which one to use.