Avoid using `ssh -A` in manual ssh instructions
Agent forwarding is bad practice, as it exposed your ssh private key to
the jumpbox (and potentially other users on that machine).

This updates the instructions to use ssh's jumpbox support instead.

Also updates the commit to find the jumpbox to a more up to date version
and fixes the app_console command (publishing-api doesn't live on the
backend machine class anymore, but publisher does).