Remove recursive rendering to simplify code
On Ben Thorner's suggestion [1], if we stop worrying about
rendering documents in nested lists and instead just list
directories and contents in turn (so no list is indented more
than twice), the code can be significantly simplified.

As a nice side-effect, we no longer need the recursion on the
_imported_docs.html.erb partial, which means we can remove the
`uls_on_newlines` hack.

[1]: https://github.com/alphagov/govuk-developer-docs/pull/3047#discussion_r610448314