Rename 'add-a-pingdom-check' to 'pingdom'
Search results for Pingdom are currently pretty ambiguous; it's
not clear what the 'canonical' resource for Pingdom documentation
is. See [screenshot](https://user-images.githubusercontent.com/5111927/115727451-8144b280-a37b-11eb-9a1c-0f1d631c05e6.png).

Of all the Pingdom docs in this repo, the 'add-a-pingdom-check'
doc is the most exhaustive, so is a good starting point for our
canonical page. I've set up a redirect so that any bookmarks or
existing links continue to work.

The generically titled 'Pingdom' seems more appropriate than the
previous, as the document also goes into detail about what Pingdom
is, how to access it, and when to add a check - not JUST how to
add the check.