Rewrite Pingdom sections
The "When to add a Pingdom check" section duplicated some of the
information in the introduction, and was not particularly clearly
worded. I've now simplified and shortened this as an opening
paragraph to the (renamed) "Add a Pingdom check" section.

I've highlighted the difference between internal and external
monitoring in the intro. And I've made the use of inline links
consistent throughout the file, which was using a mixture of inline
and alias links.