Remove references to gds vpn command
The VPN command in the GDS CLI no longer works as the GDS VPN uses
Google SSO to authenticate. The instructions that remain aren't current
up to date, but I'm not sure where to link to instead so I've left them
for now and we can update them in the future.