Add how to create a local transaction documentation
The Covid-19 product team have been asked to implement a few new local transaction
services over the last few months.

As the majority of the team have not worked with local links manager or local transactions
before, we want to be sure that we have everything we need before implementing any new
local transaction services.

Unfortunately, the steps required to create a local transaction service do not appear to
be well documented.

This change adds the missing documentation.