Make instructions consistent and clearer between Int & Staging
I ended up signing up for subscriptions to the Staging Google account on Integration (by accident), then on Production (on purpose), then on Staging. Only the latter came through with a confirmation email. This makes sense on reflection, but I think can be made a little clearer from the documentation.