Rename 'master' to 'main'
This commit renames all references from 'master' to 'main', where:

- the repo in question has changed its default branch to 'main'
- the docs in question are generic and can be updated to refer to
  'main' without breaking the effectiveness of the docs

In a few places I've changed the URL altogether, where it
referenced 'master' in the URL and both the 'master' and 'main'
variants led to 404s, indicating the files have been moved or
deleted. In these cases, I've tried to link to the commit in which
it was deleted, or the location of the new file.