Add Special Route Publisher to list of applications
This adds in a somewhat strange utility, Special Route Publisher, which
is used as a means to publish routes for content on GOV.UK via a rake
task interface. It is not a traditional application as it doesn't run
continuously, instead being executed on demand as part of a jenkins
task. However, it is still known as a publishing application via GOV.UK
metadata which describes it as the "publishing_app" for content it has
published.

I felt that this tool belonged in this list as this tool is a distinct part
of the GOV.UK publishing stack.