Don't show a deploy URL for apps without production_hosting
This stops apps that aren't deployed from showing links to deploy
scripts. Previously these apps all showed incorrect links to
govuk_app_deployments.

The only "app" here that did have a deploy script that could be linked
to is govuk-puppet, which has a deploy script in govuk-secrets [1]. I
did try to link to this however the deploy_url option is rather complex
where setting it only takes affect for an app hosted on the PASS. Rather
than fix this I reasoned that the puppet deploy link is on a private
repo.

[1]: https://github.com/alphagov/govuk-secrets/blob/main/puppet_aws/deploy.sh