Add 5xx example Athena query
Add an example AWS Athena query to give the total number of 5xx responses,
the total number of requests, and the percentage of 5xx responses of the
total number of requests.

In cases where we have high 5xx responses, it saves you having to work
out the total and % manually.