Document dpkg issue
A few times recently, we've seen Puppet failures due to underlying
dependencies being unfulfilled. This appears to be down to the
init scripts being interrupted by unattended reboots.

We already have this phenomenon documented, but don't signpost it
very well or describe how to fix it. This commit is an attempt to
address that.