Signpost the reprovisioning docs a bit more
Even though I documented this yesterday, I struggled to find the
doc by searching, and it wasn't linked to from the "check process
running" doc. I knew the underlying cause but couldn't remember
the fix.

This improves the signposting to the doc, and also clarifies that
the workaround requires `sudo`.