Document how to deploy Fastly from our individual accounts
We no longer have a shared 2nd line account for Fastly, so can
no longer rely on a centrally stored API key for deployments. This
PR documents how to create an access token linked to your own
Fastly account, and how to use it to deploy the CDN configuration
for `assets` and `www`. (There is also a third option,
`performanceplatform`, but as this is retired, I assume it should
be removed and we've no need to document it here).

Trello: https://trello.com/c/ShvEW3om/2560-add-documentation-on-how-to-use-individual-fastly-accounts-2