Provide links to Jenkins for Elasticsearch Rake tasks
There are a number of Rake tasks listed in the Elasticsearch reindexing
documentation that suggest running the task directly on a machine.

This is generally discouraged, with Rake tasks preferred to be run as a
Jenkins job.

Therefore updating the documentation to provide links to Jenkins for
each environment, which also reduces the amount of copy and pasting
needed.