Describe the process for rebooting docker-management machines
The previous instructions were open to ambiguity. For example,
I interpeted them as "disable the unattended reboots on the
docker_management machine class", but this made little sense as
we only had one such class on Production, so why would it be
necessary? It's only by digging into an old govuk-puppet [commit]
that it's clear that actually we need to disable unattended reboots
in ALL machines.

I've also removed the unhelpful link to
https://github.com/alphagov/govuk-puppet/blob/master/hieradata_aws/class/docker_management.yaml,
as this doesn't provide any additional information that isn't
already in the doc.

[commit]: https://github.com/alphagov/govuk-puppet/commit/97ab27fc9a2d79305111280ee3f2a74a0d0f9e12