Update DNS docs to notify senior tech and CDDO
CDDO colleagues monitor the DNS records for gov.uk, and they would like
us to notify them before making changes so they know to ignore the
alarms we set off.

We also need to tell the GOV.UK lead technologists because they're the
people who'll get a call from JISC to confirm. And also it's generally
good for them to know what's going on.