Break up applications.yml into new 'repos/' folder
Not every repo on the list is an application, so this was
confusing naming. The config file was very long and difficult to
navigate.

I've generally grouped the repos according to the 'heading comment'
they came under in the applications.yml file, e.g.
"# Frontend apps", although I have tidied up one or two (e.g.
'govuk-attribute-service-prototype' is now retired, so I've moved
that into the retired.yml file.

I've organised each file alphabetically, except for the retired.yml
file, where there appears to be roughly a chronological order of
retirements, which seems valuable to keep. I did also consider
moving the retired.yml repos back into whatever respective config
file they would have been in had they not been retired (as we
still denote their retired status with the `retired: true` config),
but also see the appeal in keeping the other files free of legacy
information, so have left as-is for now.