Minimal changes to fix build after changing data structures
Fix the Applications class to pull in data from all files under
`/data/repos`, rather than `/data/applications.yml`. It's not
ideal that the class itself is still called Applications, given
that not every repo listed is an application, but this refactor
doesn't worsen the situation, as this has always been the case.