Remove 'None' option for hosted apps
'Hosted on: None' is a confusing concept ('where is it
hosted, then?'). Looking at the list of 'applications'
(sic) currently marked as hosted on 'none', most of
them aren't applications in the strict sense of the
word. E.g. govuk_schemas is a gem, used by apps, but
not an app in itself.

Removing the 'none' option here removes a row from the
'Hosting' table here:
https://docs.publishing.service.gov.uk/apps.html

Whilst we still want to surface links to important
repos like govuk_schemas, they should not be surfaced
in this way.