Set local GitHub page cache time to 1 week
This value is only used locally, and often local development
takes longer than an hour, meaning every page needs to be
re-fetched from GitHub when previewing locally. Better to
increase the cache time to avoid that happening; we can
always clear the cache manually if we need the very latest
data.

I've added a TODO for a startup.sh CLI option that clears
the cache automatically, which would be useful, but out of
scope for this work.