Update RabbitMQ docs to add detail about the web control panel
This commit updates the docs on RabbitMQ to add details about the web control panel, specifically how it can be used during the process of rebooting instances of the cluster to aid with monitoring when those instances rejoin the cluster. The main RabbitMQ doc page is also updated to add the username and point to where to decrypt the password within the secrets repo.