Rename 'Applications' to 'Repos'
We want to extend the scope of this to handle all GOV.UK repos,
so that we can easily import useful documentation from across
GitHub, regardless of whether or not said repo is a deployable
application.

It would be confusing to push too far down this path whilst still
retaining the name 'Applications'.